Kada pronađe lokaciju sa koje je slika uzeta treba da stoji u okviru od 100 metara od tačke 30 sekundi pre nego što se evidentira da je izazov kompletiran. Potrebno jos dodati neki vid obavestenja kada turista udje u potrebni radius i kada se izazov uspesno kompletira

+++ src/Explorer.Encounters.API/Public/IEncounterProgressService.cs
@@ -14,8 +14,8 @@ namespace Explorer.Encounters.API.Public
         EncounterDto Update(long id);
         void Delete(long id);
         bool CheckEncounterProgress(long encounterId);
-        void ActivateHiddenLocationForUser(long encounterId, int userId);
-        void OnUserLocationChanged(int userId);
+        void ActivateHiddenLocationForUser(long encounterId, long userId);
+        void OnUserLocationChanged(long userId);
         //List<int> CheckParticipantsLocation(long encounterId);
         //void FinishEncounterProgress(long encounterId, List<int> users);
     }
